backend/requirements.txt
absl-py==1.0.0
anyio==3.5.0
asgiref==3.4.1
astunparse==1.6.3
cachetools==4.2.4
certifi==2021.10.8
cfgv==3.3.1
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
cycler==0.11.0
distlib==0.3.4
fastapi==0.71.0
filelock==3.4.2
flatbuffers==2.0
fonttools==4.28.5
gast==0.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.43.0
h11==0.12.0
h5py==3.6.0
identify==2.4.1
idna==3.3
importlib-metadata==4.10.0
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==12.0.0
Markdown==3.3.6
matplotlib==3.5.1
nodeenv==1.6.0
numpy==1.22.0
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.5
Pillow==9.0.0
pip==21.2.4
platformdirs==2.4.1
pre-commit==2.16.0
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.9.0
pyparsing==3.0.6
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2021.3
PyYAML==6.0
requests==2.27.0
requests-oauthlib==1.3.0
rsa==4.8
setuptools==58.1.0
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.23.1
termcolor==1.1.0
toml==0.10.2
typing_extensions==4.0.1
urllib3==1.26.7
uvicorn==0.16.0
virtualenv==20.13.0
Werkzeug==2.0.2
wheel==0.37.1
wrapt==1.13.3
zipp==3.7.0